okay

+++ src/app/login/login.component.ts
@@ -5,6 +5,9 @@ import { StudentService } from '../student.service';
 import Swal from 'sweetalert2';
 import { CommonModule } from '@angular/common';
 import { response } from 'express';
+import { HttpClient, HttpResponse } from '@angular/common/http';
+import { firstValueFrom, Observable } from 'rxjs';
+import { HTTPResponse } from '../httpresponse';
 
 
 @Component({
@@ -15,7 +18,7 @@ import { response } from 'express';
   styleUrl: './login.component.css'
 })
 export class LoginComponent {
-  constructor(private studentservice: StudentService, private route: Router) {}
+  constructor(private studentservice: StudentService, private route: Router, private http: HttpClient) {}
 
   lrn:any;
   learner: any;
@@ -31,46 +34,47 @@ export class LoginComponent {
       this.studentservice.loginLearner(this.loginLearner.value).subscribe({
         next: (response: any ) => {
           const token = response.token;
-          console.log('Token:', token);
-          localStorage.setItem('authToken', token);
-          console.log('Login Successful', response)
-          
-          this.studentservice.getLearnerByToken(token).subscribe({
-                next: (data) => {
-                  this.learner = data;
-                  const lrn = data.lrn; // Adjust based on your API response structure
-                  if (lrn) {
-                    localStorage.setItem('LRN', lrn);
-                    this.lrn = lrn;
-                    Swal.fire({
-                      position: "center",
-                      icon: "success",
-                      title: "You are now logged in!",
-                      showConfirmButton: false,
-                      timer: 1000
-                    });
-                    this.route.navigate(['/main/Home']);
+          if(response.token == null) {
+            Swal.fire({
+              position: "center",
+              icon: "warning",
+              title: "Provided Credentials are Incorrect",
+              showConfirmButton: false,
+              timer: 3000
+            });
+          } else {            
+            console.log('Token:', token);
+            localStorage.setItem('authToken', token);
+            console.log('Login Successful', response)
+            
+            this.studentservice.getLearnerByToken(token).subscribe({
+                  next: (data) => {
+                    this.learner = data;
+                    const lrn = data.lrn; // Adjust based on your API response structure
+                    if (lrn) {
+                      localStorage.setItem('LRN', lrn);
+                      this.lrn = lrn;
+                      Swal.fire({
+                        position: "center",
+                        icon: "success",
+                        title: "You are now logged in!",
+                        showConfirmButton: false,
+                        timer: 1000
+                      });
+                      this.route.navigate(['/main/Home']);
+                    }
                   }
-                }
-          })
+            })
+          }
         },
-        error: (error) => {
-          console.error('Login', error);
-          if(error === 'Bad Request') {
+        error: (error: any) => {
+          if(error === 'Bad Request' && error.response == null) {
             Swal.fire({
               position: "center",
               icon: "warning",
               title: "You are not enrolled yet! (LRN Needed)",
               showConfirmButton: false,
-              timer: 1000
-            });
-          } else if(error === 'Unauthorized') {
-            Swal.fire({
-              position: "center",
-              icon: "warning",
-              title: "The Provided Credentials are incorrect",
-              showConfirmButton: false,
-              timer: 1000
+              timer: 3000
             });
           }
         }
@@ -83,7 +87,18 @@ export class LoginComponent {
       return;
     }
   }
-  
+
+  // onSubmit() {
+  //   this.loginService(this.loginLearner.value)
+  //   .subscribe({
+  //     next: (data) => console.log(data.message),
+  //     error: (err) => console.log(err)
+  //   })
+  // }
+
+  loginService(data: any): Observable<HTTPResponse> {
+    return this.http.post<HTTPResponse>('http://localhost:8000/api/loginLearner', data)
+  }
   
 
   togglePasswordVisibility(): void {
